requirements.txt
flask
clip-openai
torch
faiss-cpu
numpy
Pillow
requests
pandas
# Scraping için (gerekliyse):
# selenium
# beautifulsoup4